feat: added dialog confirm window

+++ src/pages/Dashboard/Dashboard.jsx
@@ -8,15 +8,53 @@ import { LoadingSpinner } from '../../components/LoadingSpinner'
 import { NoPostFound } from '../../components/NoPostFound'
 import { PostCard } from '../../components/PostCard'
 
+import Swal from 'sweetalert2'
+import withReactContent from 'sweetalert2-react-content'
+
 export const Dashboard = () => {
+  const swal = withReactContent(Swal)
+
   const { user } = useAuthValue()
   const uid = user.uid
   // const posts = []
   const { posts, loading } = useFetchPosts('posts', null, uid)
   const { deletePost } = userDeletePost('posts')
 
+  const swalert = (id) => {
+    swal.fire({
+      title: 'Tem certeza disso?',
+      text: "Não será possível voltar atrás!",
+      icon: 'warning',
+      color: '#fff',
+      background: '#2A2634',
+      backdrop: `
+        rgb(42 38 52 / 0.4)
+        no-repeat
+      `,
+      showCancelButton: true,
+      cancelButtonText: 'Cancelar',
+      confirmButtonColor: '#3B71FE',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Sim, pode apagar!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        deletePost(id)
+        Swal.fire(
+          'Apagado!',
+          'Seu post foi apagado.',
+          'success'
+        )
+      }
+    })
+  }
+
   const handleDeletePost = id => {
-    deletePost(id)
+    try {
+      swalert(id)
+      // deletePost(id)
+    } catch (error) {
+      console.log(error)
+    }
   }
   return (
     <motion.main
@@ -43,14 +81,14 @@ export const Dashboard = () => {
               <>
                 {posts
                   ? posts.map(post => (
-                      <PostCard
-                        key={post.id}
-                        {...post}
-                        adminTools={true}
-                        onClick={() => handleDeletePost(post.id)}
-                        editPage={`/posts/edit/${post.id}`}
-                      />
-                    ))
+                    <PostCard
+                      key={post.id}
+                      {...post}
+                      adminTools={true}
+                      onClick={() => handleDeletePost(post.id)}
+                      editPage={`/posts/edit/${post.id}`}
+                    />
+                  ))
                   : null}
               </>
             )}
